refactor: split options unit tests into focused modules
